+++ app/controllers/variables_controller.rb
@@ -1,8 +1,9 @@
 class VariablesController < ApplicationController
-  before_action :set_viewable_dataset,      only: [:show, :index, :image]
-  before_action :redirect_without_dataset,  only: [:show, :index, :image]
-  before_action :set_viewable_variable,     only: [:show, :image]
-  before_action :redirect_without_variable, only: [:show]
+  before_action :set_viewable_dataset,      only: [:show, :graphs, :known_issues, :related, :history, :index, :image]
+  before_action :redirect_without_dataset,  only: [:show, :graphs, :known_issues, :related, :history, :index, :image]
+  before_action :set_viewable_variable,     only: [:show, :graphs, :known_issues, :related, :history, :image]
+  before_action :redirect_without_variable, only: [:show, :graphs, :known_issues, :related, :history]
+  before_action :set_folder_and_version,    only: [:show, :graphs, :known_issues, :related, :history]
 
   def index
     variable_scope = @dataset.variables.search(params[:s]).with_folder(params[:folder])
@@ -19,13 +20,18 @@ class VariablesController < ApplicationController
   end
 
   def show
-    graphs_folder = File.join(@dataset.root_folder, 'dd', 'graphs')
-    valid_folders = if File.exist?(graphs_folder)
-                      (Dir.entries(graphs_folder) - ['.','..']).select { |f| !File.file?(f) }
-                    else
-                      []
-                    end
-    @version = (valid_folders.include?(params[:v].to_s.strip) ? valid_folders.select{|f| f == params[:v].to_s.strip}.first : @variable.version.to_s)
+  end
+
+  def graphs
+  end
+
+  def related
+  end
+
+  def known_issues
+  end
+
+  def history
   end
 
   def image
@@ -50,5 +56,15 @@ class VariablesController < ApplicationController
 
   def redirect_without_variable
     empty_response_or_root_path(dataset_variables_path(@dataset)) unless @variable
+  end
+
+  def set_folder_and_version
+    graphs_folder = File.join(@dataset.root_folder, 'dd', 'graphs')
+    @valid_folders = if File.exist?(graphs_folder)
+                      (Dir.entries(graphs_folder) - ['.','..']).select { |f| !File.file?(f) }
+                    else
+                      []
+                    end
+    @version = (@valid_folders.include?(params[:v].to_s.strip) ? @valid_folders.select { |f| f == params[:v].to_s.strip }.first : @variable.version.to_s)
   end
 end